AgileAdapterBuilder package refactoring
