+++ src/main/java/com/agileandroid/ui/adapters/AgileAdapterBuilder.java
@@ -1,4 +1,4 @@
-package com.agileandroid.ui.adapters.recycler.builder;
+package com.agileandroid.ui.adapters;
 
 import android.content.Context;
