+++ src/main/java/com/agileandroid/ui/adapters/view/interactor/Interactor.java
@@ -1,6 +1,6 @@
 package com.agileandroid.ui.adapters.view.interactor;
 
-import com.agileandroid.ui.adapters.recycler.builder.AgileAdapterBuilder;
+import com.agileandroid.ui.adapters.AgileAdapterBuilder;
 import com.agileandroid.ui.adapters.recycler.holder.HolderRecycler;
 
 /**
