+++ src/main/java/com/agileandroid/ui/adapters/view/resolver/ViewResolver.java
@@ -3,7 +3,7 @@ package com.agileandroid.ui.adapters.view.resolver;
 import android.view.LayoutInflater;
 import android.view.View;
 
-import com.agileandroid.ui.adapters.recycler.builder.AgileAdapterBuilder;
+import com.agileandroid.ui.adapters.AgileAdapterBuilder;
 import com.agileandroid.ui.adapters.recycler.holder.HolderRecycler;
 import com.agileandroid.ui.adapters.view.interactor.Interactor;
 import com.agileandroid.ui.adapters.view.populator.Populator;
